Update messenger file name
파일 이름을 수정, 메시지 텍스트는 수신,송신 구분을 변수로 처리하여 하나의 파일로 병합
